added util for getting email and removed email from requests
