+++ frontend/src/routes/api/groups/delete/+server.ts
@@ -1,18 +1,10 @@
 import type { RequestHandler } from './$types';
 import * as db from '$lib/server/database';
-import { _getUserInfo } from '../../oauth/user/+server';
+import { getEmail } from '$lib/utils/getEmail';
 
 export const POST: RequestHandler = async ({ request, cookies }) => {
 	const sessionCookie = cookies.get('user_session');
-	let user_email = '';
-	if (sessionCookie) {
-		const sessionCookieJson = JSON.parse(sessionCookie);
-		const userData = await _getUserInfo(
-			sessionCookieJson.oauth_token,
-			sessionCookieJson.oauth_token_secret
-		);
-		user_email = userData.email;
-	}
+	const user_email = await getEmail(sessionCookie ?? '');
 	const { name } = await request.json();
 	return db.deleteUserGroup(user_email, name);
 };
