+++ frontend/src/routes/api/surveys/create/+server.ts
@@ -1,19 +1,11 @@
 import type { RequestHandler } from './$types';
 import * as db from '$lib/server/database';
 import type SurveyCreateInfo from '$lib/entities/surveys/SurveyCreateInfo';
-import { _getUserInfo } from '../../oauth/user/+server';
+import { getEmail } from '$lib/utils/getEmail';
 
 export const POST: RequestHandler = async ({ request, cookies }) => {
 	const sessionCookie = cookies.get('user_session');
-	let user_email = '';
-	if (sessionCookie) {
-		const sessionCookieJson = JSON.parse(sessionCookie);
-		const userData = await _getUserInfo(
-			sessionCookieJson.oauth_token,
-			sessionCookieJson.oauth_token_secret
-		);
-		user_email = userData.email;
-	}
+	const user_email = await getEmail(sessionCookie ?? '');
 	const info: SurveyCreateInfo = await request.json();
 	info.user_email = user_email;
 	return db.createSurvey(info);
